Clean-up ViT network definition files.
